+++ lib/gltf/parse.types.ts
@@ -1,0 +1,1 @@
+export type GlTFResolveUriFunc = (uri: string) => Promise<ArrayBufferView>;
