+++ lib/gltf/primitive.types.ts
@@ -1,0 +1,3 @@
+export type GlTF_PRIMITIVE_MODE = (
+    | 'points' | 'lines' | 'line_loop' | 'line_strip' | 'triangles' | 'triangle_strip' | 'triangle_fan'
+);
