+++ lib/gltf/schema.types.ts
